+++ lib/src/feature_discovery.dart
@@ -91,12 +91,18 @@ class DescribedFeatureOverlay extends StatefulWidget {
   @Deprecated("Replaced by backgroundColor")
   final Color color;
 
+  /// The color of the outside layout, behind texts.
+  /// This parameter replaces the [color] parameter: you should only specify this one.
   /// If null, default to [ThemeData.primaryColor]
   final Color backgroundColor;
 
   /// If null, default to current [IconTheme]
   final Color iconColor;
+
+  /// Color of the target, that is the small circle behind the icon.
   final Color targetColor;
+
+  /// Color for title and text
   final Color textColor;
   final String title;
   final String description;
@@ -140,7 +146,12 @@ class DescribedFeatureOverlay extends StatefulWidget {
         assert(enablePulsingAnimation != null),
         assert(targetColor != null),
         assert(textColor != null),
+<<<<<<< HEAD
         assert(color == null || backgroundColor == null),
+=======
+        assert(color == null || backgroundColor == null, 
+          "[color] parameter has been replaced by [backgroundColor]: as they're the same, you should only specify one"),
+>>>>>>> bb207310eae7a7f3336a45add62be9d4144585e9
         super(key: key);
 
   @override